Fix doc builds
This imperfect fix lets the docs continue building once again.

Near as I can tell, if you try to call `autofunction` on a
`functools.lru_cache` wrapped function, Sphinx cannot properly detect
that it is a function since its now technically a class?

Either way, you get this error which blocks the build:

```
AttributeError: 'functools._lru_cache_wrapper' object has no attribute '__code__'
```

If you try to `autoclass` it however, it works fine, but the rendered
automatic documentation is a bit off. So for now I added a note about
the discrepancy, made the imperfect fix, and moved on.

This also appears to be [a known problem](https://groups.google.com/forum/#!topic/dev-python/IpweOZNwRg0), however, this is also beyond
the scope of things for us to fix.